+++ src/component/Project/Project.tsx
@@ -1,17 +1,9 @@
 import Section from "../Section/Section";
 import { projects } from "./ProjectList";
 import styles from "./Project.module.css";
-import Button from "../Button/Button";
-import video from "../../assets/video.svg";
-import close from "../../assets/close.svg";
-import { useState } from "react";
+import VideoDialog from "./VideoDialog";
 
 const Project = () => {
-  const [isOpen, setIsOpen] = useState(false);
-  function handleVideoOpen() {
-    setIsOpen((prev) => !prev);
-    console.log("djc");
-  }
   return (
     <Section heading="Projects" id="project">
       <ul className={styles.porjects}>
@@ -26,19 +18,7 @@ const Project = () => {
               ))}
             </ul>
             {project.video && (
-              <>
-                <dialog open={isOpen} className={styles.videoDialog}>
-                  <video controls>
-                    <source src={project.video} type="video/mp4" />
-                  </video>
-                  <Button onClick={handleVideoOpen} variant="transparent-close">
-                    <img className={styles.close} src={close} alt="video" />
-                  </Button>
-                </dialog>
-                <Button onClick={handleVideoOpen} variant="transparent">
-                  <img className={styles.videoImg} src={video} alt="video" />
-                </Button>
-              </>
+              <VideoDialog project={project}/>
             )}
             <a className={styles.link} href={project.link}>
               GitHub
